requirements.txt
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
django-cors-headers==3.13.0
Pillow==9.3.0
requests==2.28.1
requests-toolbelt==0.10.1
cos-python-sdk-v5==1.9.22